Improvements to Market Share by dApps reporting
Two main improvements are added in this commit:

* Providing a date range for records to include in the market share report
This implements the same functionality as employed for other reports. 
The changes can mainly be found in:
blockOps.js: reportComments: Function to provide parameters for reportCommentsMongo

* Capturing the number of authors for each application via a double grouping
Records are first grouped by application and author and then by application.
This allows a count across distinct authors within each application. 
Posts are counted in the first grouping then the totals for each author summed in the second grouping.
Payment metrics are summed in each grouping.

The changes can mainly be found in:
mongoblock.js: reportCommentsMongo