Votes, active votes and curator rewards
Blockchain data on votes can be found in (at least) three places. The block transaction "vote" which is included in a block when the vote is made, the virtual operation "curation_reward" which is attached to blocks at post payout, and the aggregated vote data for a post which is updated throughout a post's life and can be extracted using the "get_active_votes" call. 

This commit adds (or, in the case of "curation_reward", reworks) the functions to extract this data from the blockchain and add it to the MongoDB. The vote data is stored in each MongoDB comment document as an array of objects, with the fields of the objects drawn from the three different sources. The "get_active_votes" call is implemented at the point of payout (past this point any updates are irrelevant).

steemrequest.js
* Addition of getActiveVotes function to return blockchain "active votes" information from a single comment defined by author / permlink. 

mongoblock.js
Addition (or reworking) of functions to process blockchain data and insert it in the MongoDB for:
* vote data (processVote, mongoVote) 
* active vote data (processActiveVote, mongoActiveVote)
* curation rewards (processCuratorReward, mongoCuratorReward)

blockOps.js
* Addition of function activeVotes to handle extraction of "active vote" data from blockchain and inclusion in MongoDB.