Adding analysis of curator rewards
This commit adds the functionality to analyse curator rewards through the curator_reward vests to voter rshares ratio (i.e. curation reward / vote size). 

The analysis calculates the ratio for all curator rewards in a chosen date range and then sorts them in order. The top 10 are output in the console.log, although all result are available if desired.

The analysis can be carried out across all accounts, or for a single specified account (entered as the third parameter in the command line).

The main changes are:

mongoblock.js
* findCuratorMongo: Function to analyse curator rewards: vests to rshares ratio for a single voter or all voters