Further robustness and re-processing
Further updates to increase robustness of processing of blocks. 

Main updates in mongoblock.js
* function validateComments
There are difficulties in separating original comments (where comments includes both posts and comments) and comment edits. This is due to the tool being applied to blocks in any order - a necessity for an analysis tool. Original 'comment' operations in blocks and operations which edit those comments appear the same, with no clear marker to distinguish them. Options to get around this issue are: 
(1) Pull the comment itself using get_content and check the created date. This results in pulling much of the comment data twice (in the block operations API call and in the get_comment call) and results in data limits from the node being exceeded. Not an option.
(2) Always run the tool forwards starting from inception or from a point in time 7 days before the analysis period commences. Not in keeping with the idea of the tool.
(3) Self-validation. The data, if run over a long enough period, or if enough data is collected, is self-validating. Author_payout operations can be used to validate whether a comment is the original or an edit, based on the time between the comment and author payout timestamps.
The validateComments function performs this task (or at least makes a first version of it - more needs to be added).

* function resetBlocksProcessed
This function resets the "operations processed" count and the notHandled record in the blocksProcessed collection so that rerun blocks (i.e. past errors or incomplete processed blocks) can be rerun.

* insertion of records into mongo (e.g. mongoComment, mongoBenefactorReward)
Updates to make the insertion process more robust, using searches of mongo database to check whether previous updates  have been made.