Implementing fx tables between steem currencies
The virtual operations of each block show the payouts for authors, curators and benefactors. These are listed in a mixture of SBD, Steem and Vests (for SP). The active vote data uses rshares to measure the vote size. Most analyses will want the results in one currency, most likely STU since this is the measure most commonly displayed by the Steem dApps.

A problem arises in that the blockchain API give no access to historic fx rates between these different measures. Vests could be translated to Steem using a formula but the Steem to STU fx rate is a rolling 3.5 day median average of the witness feed prices, meaning that all the witness price feed information would need to be collected for a period 3.5 days before the start of any analysis. There is no obvious way to translate rshares to vote values.

The payout information in STU is available from each "post" data using the get_content API. But pulling the data for every single post as well as all the block data is not an option.

As such I have implemented an "empirical" solution. This chooses a single suitable post for each hour (according to various criteria), pulls that single post using get_content, and uses the payout data to derive fx rates. For example the curator payout in STU from the post can be compared to the total vests from the curator_reward payouts from the virtual block operations to give a Vests-STU fx rate for each hour.

A similar approach is applied for Steem and rshares fx rates.

blockops.js
* function fillPrices: controls the process and does the calculations

mongoblock.js
* function mongoFillPrices: chooses a post for each hour (from the 29th or 30th minute only, with curation rewards, payout as large as possible)
* function mongoPrice: adds the fx rates to collection 'prices'